requirements.txt
customtkinter
tkcalendar
Pillow
